+++ examples/simple/src/graphql/tests/holiday_type.test.ts
@@ -17,7 +17,7 @@ test("create and edit holiday", async () => {
       schema,
       args: {
         label: "Inauguration",
-        date: dt.toMillis(),
+        date: dt.toISODate(),
         dayOfWeek: "WEDNESDAY",
         dayOfWeekAlt: "WEDNESDAY",
         log: {
@@ -27,7 +27,7 @@ test("create and edit holiday", async () => {
       },
     },
     ["holiday.label", "Inauguration"],
-    ["holiday.date", dt.toUTC().toISO()],
+    ["holiday.date", "2021-01-20"],
     ["holiday.dayOfWeek", "WEDNESDAY"],
     ["holiday.dayOfWeekAlt", "WEDNESDAY"],
     [