+++ internal/schema/input/input.go
@@ -123,6 +123,7 @@ func (n *Node) AddAssocEdgeGroup(edgeGroup *AssocEdgeGroup) {
 type GlobalSchema struct {
 	ExtraEdgeFields []*Field     `json:"extraEdgeFields,omitempty"`
 	GlobalEdges     []*AssocEdge `json:"globalEdges,omitempty"`
+	EdgeIndices     []*Index     `json:"edgeIndices,omitempty"`
 	Init            bool         `json:"init,omitempty"`
 	TransformsEdges bool         `json:"transformsEdges,omitempty"`
 	GlobalFields    []*Field     `json:"globalFields,omitempty"`