+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Changelog for the docker image are [here](/docker_CHANGELOG.md).
 
 ## [Unreleased]
+- support offset in QueryDataOptions for loadCustomData/loadCustomEnts (#1895)
+- add global edge composite indices for extra edge fields (#1897)
+- fix loadCustomCount custom clause query field aliasing (#1899)
+- allow overrideAlias to be null to skip aliasing in clauses (#1901)
 
 ## [0.2.2] - 2025-02-22
 - ability to customize default ordering for an indexed edge (#1878)